+++ front/src/assets/component/All_companies.jsx
@@ -1,51 +1,20 @@
 import React from 'react';
 import '../scss/All_companies.scss';
-import { useEffect, useState } from 'react'
+import useFetch from '../hook/useFetch';
+import Pagination from './Pagination_companies';
 
 const All_contacts = () => {
-    const [data, setData] = useState([]);
-    
-    useEffect(() => {
-        fetch("https://cogip-api.pierre-mauriello.be/companies")
-            .then ((resp) => resp.json())
-            .then ((apiData) => {
-                setData(apiData);
-            });
-    }, 
-    []);
+    const {loading, error, data} = useFetch('https://cogip-api.pierre-mauriello.be/invoices')
+
+    if(loading) return <p>Loading</p>
+    if(error) return <p>Unable to load</p>
+
     return (
         <section className='section_last_invoices'>
             <h3>All companies</h3>
             <hr className='yellow_rectangle_3'/>
             <input type="text" placeholder='Search company' className='input_text'/>
-            <section className='table_last_invoices'>
-                <table>
-                    <thead> 
-                        <tr>
-                            <th>Name</th>
-                            <th>TVA</th>
-                            <th>Country</th>
-                            <th>Type</th>
-                            <th>Created at</th>
-                        </tr> 
-                    </thead>
-                    <tbody>
-                        {
-                            data.map((companie) => {
-                                return( 
-                                <tr>
-                                    <th>{companie.name}</th>
-                                    <th>{companie.tva}</th>
-                                    <th>{companie.country}</th>
-                                    <th>{companie.type_id}</th>
-                                    <th>{companie.created_at}</th>
-                                </tr>
-                                )
-                            })
-                        }
-                    </tbody>
-                </table>
-            </section>
+            <Pagination data={data}/>
         </section>
     );
 };
