+++ front/src/assets/component/All_contacts.jsx
@@ -1,49 +1,20 @@
 import React from 'react';
 import '../scss/All_contacts.scss';
-import { useEffect, useState } from 'react'
+import useFetch from '../hook/useFetch';
+import Pagination from './Pagination_invoices';
 
 const All_contacts = () => {
-    const [data, setData] = useState([]);
-    
-    useEffect(() => {
-        fetch("https://cogip-api.pierre-mauriello.be/contacts")
-            .then ((resp) => resp.json())
-            .then ((apiData) => {
-                setData(apiData);
-            });
-    }, 
-    []);
+    const {loading, error, data} = useFetch('https://cogip-api.pierre-mauriello.be/invoices')
+
+    if(loading) return <p>Loading</p>
+    if(error) return <p>Unable to load</p>
+
     return (
         <section className='section_last_invoices'>
             <h3>All contacts</h3>
             <hr className='yellow_rectangle_2'/>
             <input type="text" placeholder='Search company' className='input_text'/>
-            <section className='table_last_invoices'>
-                <table>
-                    <thead> 
-                        <tr>
-                            <th>Invoice number</th>
-                            <th>Dates due</th>
-                            <th>Company</th>
-                            <th>Created at</th>
-                        </tr> 
-                    </thead>
-                    <tbody>
-                        {
-                            data.map((companie) => {
-                                return( 
-                                <tr>
-                                    <th>{companie.id}</th>
-                                    <th>{companie.company_id}</th>
-                                    <th>{companie.created_at}</th>
-                                    <th>{companie.updated_at}</th>
-                                </tr>
-                                )
-                            })
-                        }
-                    </tbody>
-                </table>
-            </section>
+            <Pagination data={data}/>
         </section>
     );
 };
